+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## Unreleased
+
+* Added partial translations into Polish and French.
+  (Thanks to contributors on the [Crowdin project](https://crowdin.com/project/madamiru))
+
 ## v0.1.0 (2024-12-12)
 
 * Initial release.